+++ rs/pocket_ic_server/tests/test.rs
@@ -1448,8 +1448,12 @@ fn auto_progress() {
 
     let t0 = pic.get_time();
 
+    assert!(!pic.auto_progress_enabled());
+
     // Starting auto progress on the IC => a corresponding log should be made and time should start incresing automatically now.
     pic.auto_progress();
+
+    assert!(pic.auto_progress_enabled());
 
     loop {
         let mut bytes = [0; 1000];
@@ -1471,6 +1475,8 @@ fn auto_progress() {
 
     // Stopping auto progress on the IC => a corresponding log should be made.
     pic.stop_progress();
+
+    assert!(!pic.auto_progress_enabled());
 
     loop {
         let mut bytes = [0; 1000];